Upgraded Version to 0.41 Fixed bug in calculation of Beta.μ
Added explanatory details to UnivariateHistogram[T].toString:
Histogram: {
	[ 0.00, 0.10 ) 🌑 ▕    ∝ 0.0013
	[ 0.10, 0.20 ) 🌑 ▕█▉   ∝ 0.0176
	[ 0.20, 0.30 ) 🌑 ▕█████████▍   ∝ 0.0852
	[ 0.30, 0.40 ) 🌒 ▕██████████████████    ∝ 0.1634
	[ 0.40, 0.50 ) 🌓 ▕██████████████████████████▏   ∝ 0.237
	[ 0.50, 0.60 ) 🌓 ▕█████████████████████████▋   ∝ 0.232
	[ 0.60, 0.70 ) 🌔 ▕██████████████████▉   ∝ 0.1703
	[ 0.70, 0.80 ) 🌔 ▕████████▎   ∝ 0.0755
	[ 0.80, 0.90 ) 🌔 ▕█▊   ∝ 0.0168
	[ 0.90, 1.00 ] 🌕 ︙    ∝ 9.0E-4
}

[ 0.40, 0.50 )  // encodes the bin as an interval [inclusive, exclusive)
[ 0.90, 1.00 ] // encodes the bin as an interval [inclusive, inclusive]
🌓  // separates interval label from bar and quantity.  Also visually encodes the percentage of the cumulative distribution at this bin.
▕███████▊ // visualizes the 'mass' in this bin
 ∝ 0.0709 // the 'proportion' of the total mass represented by this bin.

 Updated Verify to reflect histogram updates.

 addressed deprecation warnings about non-explicit Long to Double conversions in Poisson distributions.

+++ vector/shared/src/main/scala/ai/dragonfly/math/stats/probability/distributions/Poisson.scala
@@ -18,7 +18,7 @@ case class Poisson(λ:Double) extends DiscreteProbabilityDistribution {
   override val `σ²`:Double = λ
   lazy val σ:Double = Math.sqrt(λ)
 
-  def p(x:Long):Double = Math.exp( x * Math.log(λ) - λ - Math.log(Γ(x+1)) )
+  def p(x:Long):Double = Math.exp( x.toDouble * Math.log(λ) - λ - Math.log(Γ(x.toDouble+1.0)) )
 
   // Knuth's method:
   override def random(): Long = {
